[TASK] Merge comparison and copy actions
The copy action redirects to the compare action. The custom
template has been removed.

The button for the make action has been removed from the
compare & copy template and vice versa. This is preliminary
work for the introduction of specific filter form elements
in both templates.

Added the backend module's internal message queue which will be
rendered in a different position than the flash message queue.
The message queue survives action redirects.

Display of the "Make" and "Compare & Copy" tiles on the
welcome page.