[BUGFIX] Handle missing main frame
When the runner navigates directly to the url of a content frame,
there is no main frame with module menu etc.. To calculate the
total page size, the runner normally switches between the two frames,
which would fail in this case. The runner should check for the main
frame first before switching.