[TASK] Use chrome in headless mode
Chrome without headless mode rendered a shifted option layer when
a select box was clicked.

As a side effect of the switch to the headless mode, the browser
page size has increased: Resizing the window to e.g. 1024x768 now
results in screenshots of 1024x768 where it was previously 1016x639.

See https://github.com/php-webdriver/php-webdriver/issues/905
for further details on the select box rendering issue.