[FEATURE] Handle screenshot of partially scrolled out element
If the element is in a scrolling layer and is only partially visible,
chrome driver + chrome currently renders a broken element screenshot.
However, it works fine if you take a screenshot of a fully scrolled-in
or fully scrolled-out element. To work around this issue, the browser
window can be resized to full page size using the new action
Typo3Navigation::resizeToFullPage() before taking the screenshot of the element.

See https://github.com/php-webdriver/php-webdriver/issues/909
for further details on the element screenshot rendering issue.