[TASK] Run code coverage and create badge by GitHub Action
Run it only for the "main" branch, because the badge inclusion in
README.rst cannot take into account the current branch and always
displays the badge of the last code coverage run of any branch:
It would be irritating to see the results of branch "my-new-feature"
in branch "main", for example. Therefore, limit the creation of
code coverage badges to the "main" branch.