[FEATURE] Support for scrolling the modal dialog frame
The modal dialog in the TYPO3 backend either contains
the content directly in its body or in an Iframe of its
body. Handle both scenarios seamlessly.

Additionally support for closing the modal dialog by
new action `closeModalDialog`.

Furthermore switch explicitly to the modal dialog frame with the
new `switchToModalDialogFrame` action, e.g. to select elements
in it. Switch back to the main frame with the familiar
`switchToMainFrame` before continuing outside the modal dialog
frame.