[BUGFIX] Fixes opening a page tree with more than 1 level
The `AbstractPageTree::openTreePath` action tries to open
the child element without waiting for all children to be
loaded via AJAX.

This fix temporarily solves the problem by inserting a
fixed pause of 0.1s after opening a tree node - but this
is not an intelligent solution.

See issue
https://github.com/TYPO3/testing-framework/issues/245
for more details.