[BUGFIX] Fix exception "Package artifact not found." (#1629819799)
The artifact file builder has been introduced with TYPO3 v11.4 and
creates the artifact file on the CLI and reads the file per CGI. To
determine the file path, it uses the environment variable
`TYPO3_PATH_APP` which was distinct for CLI (=/var/www/html) and
CGI (=/var/www/html/public). With aligning both, the file gets
written to and read from the same place now, and the exception is
not thrown any longer.