[TASK] Switch local package constraint from "dev-main" to "@dev"
The "@dev" restriction is correct here, since the latest package
version of the current branch should be used instead of the
package version of the "main" branch. Even if the composer.lock
contains names of outdated branches, this should not affect the
installation of local packages.

Note: If you still want to have "dev-main" in the composer.lock,
for aesthetic reasons, use
```
COMPOSER_ROOT_VERSION=dev-main composer update
```