[TASK] Remove deep linking to table
The actions

- `setNavigationDefaultPid`
- `goToTable`
- `makeScreenshotOfTable`

were provided to navigate directly to a single view of a record table -
using the full browser window width and hiding the main frame
components like the TYO3 backend header and the module menu. This is no
longer supported by the TYPO3 backend, as the table view clipboard
module misses Javascript declarations of the main frame.

Therefore, this shortcut no longer provides any additional benefit and
will be removed. Use the usual actions to navigate to the record table
via list module and page tree and take a screenshot in the content
frame.