[TASK] NO MERGE: Mount `extension_builder` folder to prevent redundancy
As the Extension Builder is developed further in a separate folder at the same
time as the images of its documentation are prepared in this project,
the folder is mounted into this project to remove redundancy. It has to be
mounted and cannot be symlinked into the project as the path in the Docker
machine would be the path of the host machine and thus could not be accessed
by the Nginx webserver in the Docker machine.

NOTE: DO NOT MERGE THIS COMMIT INTO MAIN BUT SUPPORT FIGURE OUT
HOW TO BEST SUPPORT MOUNTING APPLICATIONS INTO THE SCREENSHOTS
PROJECT - OR FIND A WAY TO SUPPORT AUTOMATIC SCREENSHOTS IN THE
APPLICATION ITSELF.